refactor: sync user_id ke int64 dan cleanup tabel user di warehouse-api
- Update mapping user_id menjadi int64 agar sinkron dengan audit.proto dan Auth Service.
- Menghapus query JOIN ke tabel 'users' pada repository GetStockLogs karena tabel user sudah dihapus dari database warehouse (sekarang menggunakan sistem terpusat).
- Migrasi penuh logging user dari metode Header Response Injection ke fungsi AddLogFields.
- User info (ID, Username, Role) sekarang otomatis masuk ke dalam log group 'trace' melalui middleware Auth.
- Cleanup kode di middleware logger untuk menghapus sisa-sisa logika header internal yang sudah tidak digunakan.
- Penyesuaian gRPC call ke Audit Service untuk mendukung tipe data int64 pada field UserId.

+++ internal/entity/stock_log.go
@@ -11,6 +11,6 @@ type StockLog struct {
 	Quantity  int       `json:"quantity"`
 	Reason    string    `json:"reason"`
 	CreatedAt time.Time `json:"created_at"`
-	UserID    int       `json:"user_id,omitempty"`
+	UserID    int64     `json:"user_id,omitempty"`
 	UserName  *string   `json:"user_name,omitempty"`
 }
